+++ test/mocks/factories/blockChainDataFactory.ts
@@ -1,5 +1,6 @@
 import { Factory } from 'rosie';
 import { v4 } from 'uuid';
+import settings from '../../../src/config/settings';
 import Block, { IBlock } from '../../../src/models/Block';
 import BlockChainData, { IBlockChainData } from '../../../src/models/BlockChainData';
 import { blockFactory } from './blockFactory';
@@ -7,7 +8,7 @@ import { blockFactory } from './blockFactory';
 export const blockChainDataFactory = Factory.define('BlockChainData')
   .attr('_id', () => v4())
   .attr('chainAddress', 'CHAIN_ADDRESS')
-  .attr('chainId', 'ethereum')
+  .attr('chainId', settings.blockchain.homeChain.chainId)
   .attr('minter', 'MINTER')
   .sequence('blockId')
   .sequence('anchor');