+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Rename ForeignBlock model to BlockChainData
 - Execute migrations once before scheduling workers
+- Remove anchor, minter and chainAddress from Block collection
 
 ## [5.1.3] - 2022-09-30
 ### Changed